touch remove_quote.c
