+++ Makefile
@@ -13,6 +13,7 @@ SRCS =	src/main.c	\
 		src/lexer/add_token_to_argv.c	\
 		src/output_error/output_error.c	\
 		src/expansion/expand.c	\
+		src/expansion/remove_quote.c	\
 		src/parser/parse.c	\
 		src/redirection/redirect.c	\
 		src/pipeline/pipe.c
